Moved the contentful request in seperate component
